+++ src/index.ts
@@ -1,13 +1,19 @@
 /**
  * @since 0.6.0
  */
+import * as monadObservable from './MonadObservable'
 import * as observable from './Observable'
 import * as readerObservable from './ReaderObservable'
 import * as observableEither from './ObservableEither'
+import * as observableThese from './ObservableThese'
 import * as readerObservableEither from './ReaderObservableEither'
 import * as stateReaderObservableEither from './StateReaderObservableEither'
 
 export {
+  /**
+   * @since 0.6.12
+   */
+  monadObservable,
   /**
    * @since 0.6.0
    */
@@ -20,6 +26,10 @@ export {
    * @since 0.6.8
    */
   observableEither,
+  /**
+   * @since 0.6.12
+   */
+  observableThese,
   /**
    * @since 0.6.10
    */